refactor: refactored listingStyle options into schemaEnhancer
