+++ locales/en/LC_MESSAGES/volto.po
@@ -26,6 +26,9 @@ msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/SimpleCardTemplateOptions
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: config/Blocks/listingOptions
+#: config/Blocks/ListingOptions/simpleCardTemplate
+#: config/Blocks/listingVariations
 msgid "Aspetto"
 msgstr ""
 
@@ -108,10 +111,13 @@ msgid "Description"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/MapTemplateOptions
+#: config/Blocks/ListingOptions/mapTemplate
 msgid "Dimensione della mappa"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: components/ItaliaTheme/manage/Widgets/PathFiltersWidget
+#: config/Blocks/ListingOptions/utils
 msgid "Etichetta path filter"
 msgstr ""
 
@@ -166,12 +172,16 @@ msgid "Menu selezionato"
 msgstr "Menu selected"
 
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: components/ItaliaTheme/manage/Widgets/PathFiltersWidget
+#: config/Blocks/ListingOptions/utils
 msgid "Mostra i filtri per percorso"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Calendar/ListingSidebar
 #: components/ItaliaTheme/Blocks/Listing/Options/DefaultOptions
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: config/Blocks/ListingOptions/defaultOptions
+#: config/Blocks/ListingOptions/utils
 msgid "Mostra lo sfondo del blocco"
 msgstr ""
 
@@ -200,6 +210,8 @@ msgid "Open in a new tab"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: components/ItaliaTheme/manage/Widgets/PathFiltersWidget
+#: config/Blocks/ListingOptions/utils
 msgid "Path filter filtro"
 msgstr ""
 
@@ -209,6 +221,8 @@ msgid "Pause slider"
 msgstr "Pause"
 
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: components/ItaliaTheme/manage/Widgets/PathFiltersWidget
+#: config/Blocks/ListingOptions/utils
 msgid "Percorso path filter"
 msgstr ""
 
@@ -307,6 +321,8 @@ msgstr "Title..."
 #: components/ItaliaTheme/Blocks/Listing/Options/DefaultOptions
 #: components/ItaliaTheme/Blocks/Listing/Options/SimpleCardTemplateOptions
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: config/Blocks/ListingOptions/defaultOptions
+#: config/Blocks/listingVariations
 msgid "Titolo"
 msgstr ""
 
@@ -441,6 +457,8 @@ msgstr "Other documents"
 
 #: components/ItaliaTheme/Blocks/Listing/Options/CardWithImageTemplateOptions
 #: components/ItaliaTheme/Blocks/Listing/Options/CompleteBlockLinksTemplateTemplateOptions
+#: config/Blocks/ListingOptions/cardWithImageTemplate
+#: config/Blocks/ListingOptions/completeBlockLinksTemplate
 msgid "always_show_image"
 msgstr ""
 
@@ -563,6 +581,7 @@ msgid "biografia"
 msgstr "Biography"
 
 #: components/ItaliaTheme/Blocks/Listing/Options/BgColor
+#: config/Blocks/ListingOptions/defaultOptions
 msgid "block_bg_color"
 msgstr ""
 
@@ -748,6 +767,7 @@ msgid "descrizione_estesa"
 msgstr "Extended description"
 
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: config/Blocks/ListingOptions/utils
 msgid "detail_link_label"
 msgstr ""
 
@@ -989,6 +1009,7 @@ msgid "grid-gallery-max-items-exceeded"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: config/Blocks/ListingOptions/utils
 msgid "hide_dates"
 msgstr ""
 
@@ -1048,6 +1069,7 @@ msgid "listing_event_recurrence_label"
 msgstr "This event has multiple dates: see all"
 
 #: components/ItaliaTheme/Blocks/Listing/Options/ItemsColor
+#: config/Blocks/ListingOptions/defaultOptions
 msgid "listing_items_color"
 msgstr ""
 
@@ -1073,14 +1095,17 @@ msgid "mailto"
 msgstr "Send"
 
 #: components/ItaliaTheme/Blocks/Listing/Options/MapTemplateOptions
+#: config/Blocks/ListingOptions/mapTemplate
 msgid "map_size_large"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/MapTemplateOptions
+#: config/Blocks/ListingOptions/mapTemplate
 msgid "map_size_medium"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/MapTemplateOptions
+#: config/Blocks/ListingOptions/mapTemplate
 msgid "map_size_small"
 msgstr ""
 
@@ -1126,6 +1151,8 @@ msgstr "Alternative formats"
 
 #: components/ItaliaTheme/Blocks/Listing/Options/CardWithImageTemplateOptions
 #: components/ItaliaTheme/Blocks/Listing/Options/CompleteBlockLinksTemplateTemplateOptions
+#: config/Blocks/ListingOptions/cardWithImageTemplate
+#: config/Blocks/ListingOptions/completeBlockLinksTemplate
 msgid "natural_image_size"
 msgstr "Do not alter natural image sizes"
 
@@ -1722,44 +1749,56 @@ msgid "share"
 msgstr "Share"
 
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: config/Blocks/ListingOptions/utils
 msgid "show_description"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: config/Blocks/ListingOptions/utils
 msgid "show_detail_link"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/BandiInEvidenceTemplateOptions
+#: config/Blocks/ListingOptions/bandiInEvidenceTemplate
 msgid "show_ente"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: config/Blocks/ListingOptions/utils
 msgid "show_icon"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/MapTemplateOptions
+#: config/Blocks/ListingOptions/mapTemplate
 msgid "show_map_full_width"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/RibbonCardTemplateOptions
+#: config/Blocks/ListingOptions/ribbonCardTemplate
 msgid "show_only_first_ribbon"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: config/Blocks/ListingOptions/utils
 msgid "show_section"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/BandiInEvidenceTemplateOptions
+#: config/Blocks/ListingOptions/bandiInEvidenceTemplate
 msgid "show_tipologia"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/CardWithImageTemplateOptions
 #: components/ItaliaTheme/Blocks/Listing/Options/CompleteBlockLinksTemplateTemplateOptions
 #: components/ItaliaTheme/Blocks/Listing/Options/InEvidenceTemplateOptions
+#: config/Blocks/ListingOptions/cardWithImageTemplate
+#: config/Blocks/ListingOptions/completeBlockLinksTemplate
+#: config/Blocks/ListingOptions/inEvidenceTemplate
 msgid "show_topics"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: config/Blocks/ListingOptions/utils
 msgid "show_type"
 msgstr ""
 
@@ -1769,11 +1808,17 @@ msgstr "Silence assent"
 
 #: components/ItaliaTheme/Blocks/Listing/Options/SimpleCardTemplateOptions
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: config/Blocks/listingOptions
+#: config/Blocks/ListingOptions/simpleCardTemplate
+#: config/Blocks/listingVariations
 msgid "simplecard_listing_appearance_compact"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/Listing/Options/SimpleCardTemplateOptions
 #: components/ItaliaTheme/Blocks/Listing/Options/TemplatesOptions
+#: config/Blocks/listingOptions
+#: config/Blocks/ListingOptions/simpleCardTemplate
+#: config/Blocks/listingVariations
 msgid "simplecard_listing_appearance_description"
 msgstr ""
