+++ src/components/ItaliaTheme/index.js
@@ -14,6 +14,7 @@ export TextEditorWidget from '@italia/components/ItaliaTheme/manage/Widgets/Text
 export LinkToWidget from '@italia/components/ItaliaTheme/manage/Widgets/LinkToWidget';
 export ColorListWidget from '@italia/components/ItaliaTheme/manage/Widgets/ColorListWidget';
 export FileWidget from '@italia/components/ItaliaTheme/manage/Widgets/FileWidget';
+export PathFiltersWidget from '@italia/components/ItaliaTheme/manage/Widgets/PathFiltersWidget';
 
 /********* ICONS ********* */
 export getItemIcon from '@italia/components/ItaliaTheme/Icons/common/common';
