+++ src/config/Blocks/ListingOptions/bandiInEvidenceTemplate.js
@@ -1,0 +1,42 @@
+import { defineMessages } from 'react-intl';
+
+import { templatesOptions } from '@italia/config/Blocks/ListingOptions';
+
+const messages = defineMessages({
+  show_ente: {
+    id: 'show_ente',
+    defaultMessage: "Mostra l'ente",
+  },
+  show_tipologia: {
+    id: 'show_tipologia',
+    defaultMessage: 'Mostra la tipologia',
+  },
+});
+
+export const addBandiInEvidenceTemplateOptions = (
+  schema,
+  formData,
+  intl,
+  position = 0,
+) => {
+  let pos = position;
+
+  pos = templatesOptions(
+    schema,
+    formData,
+    intl,
+    ['show_description', 'show_ente', 'show_tipologia'],
+    {
+      show_ente: {
+        default: false,
+        label: intl.formatMessage(messages.show_ente),
+      },
+      show_tipologia: {
+        default: false,
+        label: intl.formatMessage(messages.show_tipologia),
+      },
+    },
+    pos,
+  );
+  return pos;
+};
