+++ src/config/Blocks/ListingOptions/index.js
@@ -1,0 +1,15 @@
+export {
+  addSchemaField,
+  templatesOptions,
+} from '@italia/config/Blocks/ListingOptions/utils';
+export addDefaultOptions from '@italia/config/Blocks/ListingOptions/defaultOptions';
+export {
+  addSimpleCardTemplateOptions,
+  SimpleCardTemplateAppearance_COMPACT,
+} from '@italia/config/Blocks/ListingOptions/simpleCardTemplate';
+export { addCardWithImageTemplateOptions } from '@italia/config/Blocks/ListingOptions/cardWithImageTemplate';
+export { addInEvidenceTemplateOptions } from '@italia/config/Blocks/ListingOptions/inEvidenceTemplate';
+export { addRibbonCardTemplateOptions } from '@italia/config/Blocks/ListingOptions/ribbonCardTemplate';
+export { addMapTemplateOptions } from '@italia/config/Blocks/ListingOptions/mapTemplate';
+export { addCompleteBlockLinksTemplateOptions } from '@italia/config/Blocks/ListingOptions/completeBlockLinksTemplate';
+export { addBandiInEvidenceTemplateOptions } from '@italia/config/Blocks/ListingOptions/bandiInEvidenceTemplate';
